+++ src/RcppExports.cpp
@@ -22,6 +22,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// hclust1d_heapbased
+List hclust1d_heapbased(NumericVector& points, int method);
+RcppExport SEXP _hclust1d_hclust1d_heapbased(SEXP pointsSEXP, SEXP methodSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector& >::type points(pointsSEXP);
+    Rcpp::traits::input_parameter< int >::type method(methodSEXP);
+    rcpp_result_gen = Rcpp::wrap(hclust1d_heapbased(points, method));
+    return rcpp_result_gen;
+END_RCPP
+}
 // hclust1d_single
 List hclust1d_single(NumericVector& points);
 RcppExport SEXP _hclust1d_hclust1d_single(SEXP pointsSEXP) {
@@ -33,10 +45,21 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// heap_experiments
+void heap_experiments();
+RcppExport SEXP _hclust1d_heap_experiments() {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    heap_experiments();
+    return R_NilValue;
+END_RCPP
+}
 
 static const R_CallMethodDef CallEntries[] = {
     {"_hclust1d_dedistance", (DL_FUNC) &_hclust1d_dedistance, 2},
+    {"_hclust1d_hclust1d_heapbased", (DL_FUNC) &_hclust1d_hclust1d_heapbased, 2},
     {"_hclust1d_hclust1d_single", (DL_FUNC) &_hclust1d_hclust1d_single, 1},
+    {"_hclust1d_heap_experiments", (DL_FUNC) &_hclust1d_heap_experiments, 0},
     {NULL, NULL, 0}
 };