+++ NEWS.md
@@ -1,4 +1,19 @@
 
+# hclust1d 0.0.1.9002
+
+- Added `centroid` linkage support
+
+# hclust1d 0.0.1.9001
+
+- Added `average` linkage support
+- Added a listing of all currently supported distance methods in `supported_dist.methods`
+
+# hclust1d 0.0.1.9000
+
+- Added a heap implementation of priority que with key update by id
+- Added `complete` linkage support
+- Added a listing of all currently supported distance methods in `supported_methods`
+
 # hclust1d 0.0.1
 
-- The initial version with single linkage, distance support and unit tests
+- The initial version with `single` linkage support, distance support and unit tests